fix: fi style
