+++ packages/upload/src/utils.ts
@@ -1,5 +1,44 @@
 import { isArray, isFunction } from "@illa-design/system"
 import { UploadItem, InternalDataTransferItem } from "./interface"
+import {
+  FileDefaultIcon,
+  FileMusicIcon,
+  FilePdfIcon,
+  FilePictureIcon,
+  FileVideoIcon,
+} from "@illa-design/icon"
+
+export const getIconType = (file: UploadItem) => {
+  let type = ""
+  if (file.originFile && file.originFile.type) {
+    type = file.originFile.type
+  } else {
+    const name = file.name || ""
+    const fileExtension = name.split(".").pop() || ""
+    type = fileExtension
+    if (["png", "jpg", "jpeg", "bmp", "gif"].indexOf(fileExtension) > -1) {
+      type = "image"
+    } else if (["mp4", "m2v", "mkv"].indexOf(fileExtension) > -1) {
+      type = "video"
+    } else if (["mp3", "wav", "wmv"].indexOf(fileExtension) > -1) {
+      type = "audio"
+    }
+  }
+
+  if (type.indexOf("image") > -1) {
+    return FilePictureIcon
+  }
+  if (type.indexOf("pdf") > -1) {
+    return FilePdfIcon
+  }
+  if (type.indexOf("audio") > -1) {
+    return FileMusicIcon
+  }
+  if (type.indexOf("video") > -1) {
+    return FileVideoIcon
+  }
+  return FileDefaultIcon
+}
 
 export const isAcceptFile = (
   file: File,
@@ -119,7 +158,11 @@ export const loopDirectory = (
 
 export const getFileURL = (file: UploadItem) => {
   const { originFile } = file
-  return file.url !== undefined
+  const url =
+    file.url ||
+    (file?.response && "url" in file.response && file.response.url) ||
+    undefined
+  return url !== undefined
     ? file.url
     : originFile && isFunction(URL.createObjectURL)
     ? URL.createObjectURL(originFile)
